backend/requirements.txt
flask
scikit-learn
numpy
spacy
pyspellchecker
spacy[pt]